requirements.txt
folium==0.14.0
keras==3.0.2
matplotlib==3.7.2
numpy==1.24.2
pandas==1.5.3
Pillow==9.3.0
Pillow==10.2.0
plotly==5.18.0
scikit_learn==1.3.2
seaborn==0.13.1
streamlit_folium==0.16.0
tensorflow==2.15.0.post1
